[T] Big refactoring and added documentation
The command logging is now handled in a special service that is triggered by `beforeControllerInvocation` slot.
The StandardController is separated into a UserController and CommandController, where the latter is only used for
dispatching arbitrary commands from a single entry point (for example when replaying recorded commands).

Also User is seperated into a write and read model, where the write side is only concerned with data required for
assuring business rules.